pyproject.toml
[project]
name = "shopping-agent-mcp"
version = "0.1.0"
requires-python = ">=3.10"
description = "Shopping Agent MCP Server powered by Dedalus Labs"
dependencies = [
    "dedalus-mcp>=0.7.0",
    "python-dotenv>=1.0.0",
    "duckduckgo-search>=6.0.0",
    "playwright>=1.40.0",
    "pydantic>=2.0.0",
    "requests>=2.28.0",
]
